refactor(ui): decompose dashboard pages + toast notifications
Toast system (#602):
- New ToastProvider context + useToast() hook
- Auto-dismiss after 4s, slide-in animation, success/error/info types
- Replace alert() calls in admin users page
- Wire into app layout

Today page (#576): 730 → 203 lines
- Extract BudgetRing, TokenBar, GrantCard to components/today/
- Extract fmtTokens to utils.ts
- Move 358 lines of <style jsx> to today.css

Admin users page (#568): 796 → 346 lines
- Extract CreateUserModal, BudgetModal, GrantsModal, DeleteUserModal
- Each modal owns its state, validation, and API calls
- Page keeps: user list, pagination, table, action handlers

Closes #602, #576, #568
