+++ ui/src/app/today/page.tsx
@@ -1,181 +1,13 @@
 "use client";
 
-import { useTodayBudget, type TodayModelUsage, type TodayGrant } from "@/hooks/useTodayBudget";
+import { useTodayBudget } from "@/hooks/useTodayBudget";
 import { ErrorBanner } from "@/components/ErrorBanner";
 import { useEffect, useMemo, useState } from "react";
-
-/** Format a number with k/M suffixes for large token counts. */
-function fmtTokens(n: number): string {
-  if (n >= 1_000_000) return `${(n / 1_000_000).toFixed(1)}M`;
-  if (n >= 1_000) return `${(n / 1_000).toFixed(1)}k`;
-  return n.toLocaleString();
-}
-
-/** Mini radial gauge for the hero section. */
-function BudgetRing({ percent, spent, limit, remaining }: {
-  percent: number;
-  spent: number;
-  limit: number;
-  remaining: number;
-}) {
-  const [offset, setOffset] = useState(2 * Math.PI * 45);
-  const r = 45;
-  const circumference = 2 * Math.PI * r;
-
-  useEffect(() => {
-    const progress = Math.min(percent, 100) / 100;
-    const t = setTimeout(() => setOffset(circumference - progress * circumference), 80);
-    return () => clearTimeout(t);
-  }, [percent, circumference]);
-
-  const color =
-    percent >= 100 ? "var(--error)" :
-    percent >= 80 ? "var(--warning)" :
-    "var(--accent)";
-
-  return (
-    <div className="today-ring-wrap">
-      <svg viewBox="0 0 120 120" className="today-ring-svg">
-        {/* Subtle glow */}
-        <defs>
-          <filter id="glow">
-            <feGaussianBlur stdDeviation="3" result="blur" />
-            <feMerge>
-              <feMergeNode in="blur" />
-              <feMergeNode in="SourceGraphic" />
-            </feMerge>
-          </filter>
-        </defs>
-        {/* Track */}
-        <circle cx="60" cy="60" r={r} fill="none" stroke="var(--bg-tertiary)" strokeWidth="10" />
-        {/* Progress */}
-        <circle
-          cx="60" cy="60" r={r}
-          fill="none"
-          stroke={color}
-          strokeWidth="10"
-          strokeDasharray={circumference}
-          strokeDashoffset={offset}
-          strokeLinecap="round"
-          filter="url(#glow)"
-          style={{ transition: "stroke-dashoffset 1.8s cubic-bezier(0.4, 0, 0.2, 1), stroke 0.3s" }}
-          transform="rotate(-90 60 60)"
-        />
-        <text x="60" y="55" textAnchor="middle" className="today-ring-pct" style={{ fill: color }}>
-          {Math.round(percent)}%
-        </text>
-        <text x="60" y="70" textAnchor="middle" className="today-ring-label">
-          used
-        </text>
-      </svg>
-
-      <div className="today-ring-meta">
-        <div className="today-ring-stat">
-          <span className="today-ring-stat-label">Spent</span>
-          <span className="today-ring-stat-value" style={{ color }}>${spent.toFixed(2)}</span>
-        </div>
-        <div className="today-ring-divider" />
-        <div className="today-ring-stat">
-          <span className="today-ring-stat-label">Limit</span>
-          <span className="today-ring-stat-value">${limit.toFixed(2)}</span>
-        </div>
-        <div className="today-ring-divider" />
-        <div className="today-ring-stat">
-          <span className="today-ring-stat-label">Left</span>
-          <span className="today-ring-stat-value" style={{ color: remaining <= 0 ? "var(--error)" : "var(--success)" }}>
-            ${Math.max(0, remaining).toFixed(2)}
-          </span>
-        </div>
-      </div>
-    </div>
-  );
-}
-
-/** Horizontal stacked bar for a single model's token split. */
-function TokenBar({ model }: { model: TodayModelUsage }) {
-  const total = model.inputTokens + model.outputTokens;
-  const inPct = total > 0 ? (model.inputTokens / total) * 100 : 0;
-
-  return (
-    <div className="today-model-row">
-      <div className="today-model-info">
-        <span className="today-model-name mono">{model.model}</span>
-        <span className="today-model-provider">{model.provider}</span>
-      </div>
-      <div className="today-model-metrics">
-        <span className="today-model-metric">
-          <span className="today-model-metric-label">Requests</span>
-          <span className="today-model-metric-value">{model.callCount.toLocaleString()}</span>
-        </span>
-        <span className="today-model-metric">
-          <span className="today-model-metric-label">In</span>
-          <span className="today-model-metric-value">{fmtTokens(model.inputTokens)}</span>
-        </span>
-        <span className="today-model-metric">
-          <span className="today-model-metric-label">Out</span>
-          <span className="today-model-metric-value">{fmtTokens(model.outputTokens)}</span>
-        </span>
-        <span className="today-model-metric">
-          <span className="today-model-metric-label">Cost</span>
-          <span className="today-model-metric-value">${model.costUsd.toFixed(4)}</span>
-        </span>
-      </div>
-      <div className="today-token-bar">
-        <div className="today-token-bar-in" style={{ width: `${inPct}%` }} />
-        <div className="today-token-bar-out" style={{ width: `${100 - inPct}%` }} />
-      </div>
-    </div>
-  );
-}
-
-/** Card showing a single grant's progress. */
-function GrantCard({ grant }: { grant: TodayGrant }) {
-  const pct = Math.min(grant.percentUsed, 100);
-  const color =
-    pct >= 100 ? "var(--error)" :
-    pct >= 80 ? "var(--warning)" :
-    "var(--accent)";
-
-  return (
-    <div className="today-grant-card">
-      <div className="today-grant-header">
-        <span className="today-grant-reason">{grant.reason || "Grant"}</span>
-        {grant.expiresAt && (
-          <span className="today-grant-expiry">
-            Expires {grant.expiresAt.toLocaleDateString(undefined, { month: "short", day: "numeric" })}
-          </span>
-        )}
-      </div>
-      <div className="today-grant-bar-wrap" role="progressbar"
-           aria-valuenow={pct} aria-valuemin={0} aria-valuemax={100}
-           aria-label={`${grant.reason || "Grant"}: ${pct.toFixed(0)}% used`}>
-        <div className="today-grant-bar-fill" style={{ width: `${pct}%`, background: color }} />
-      </div>
-      <div className="today-grant-stats">
-        <span className="today-grant-stat">
-          <span className="today-grant-stat-label">Used</span>
-          <span className="today-grant-stat-value">${grant.spentUsd.toFixed(2)}</span>
-        </span>
-        <span className="today-grant-stat">
-          <span className="today-grant-stat-label">Left</span>
-          <span className="today-grant-stat-value" style={{ color: grant.remainingUsd <= 0 ? "var(--error)" : "var(--success)" }}>
-            ${Math.max(0, grant.remainingUsd).toFixed(2)}
-          </span>
-        </span>
-        <span className="today-grant-stat">
-          <span className="today-grant-stat-label">Total</span>
-          <span className="today-grant-stat-value">${grant.amountUsd.toFixed(2)}</span>
-        </span>
-        {grant.grantedBy && (
-          <span className="today-grant-stat">
-            <span className="today-grant-stat-label">By</span>
-            <span className="today-grant-stat-value today-grant-by">{grant.grantedBy.split("@")[0]}</span>
-          </span>
-        )}
-      </div>
-    </div>
-  );
-}
+import { BudgetRing } from "@/components/today/BudgetRing";
+import { TokenBar } from "@/components/today/TokenBar";
+import { GrantCard } from "@/components/today/GrantCard";
+import { fmtTokens } from "@/components/today/utils";
+import "./today.css";
 
 export default function TodayPage() {
   const { data, loading, error, refresh } = useTodayBudget();
@@ -366,364 +198,6 @@ export default function TodayPage() {
         </div>
       </div>
 
-      <style jsx>{`
-        /* ── Hero Budget Section ── */
-        .today-hero {
-          background: linear-gradient(135deg, var(--bg-secondary) 0%, var(--bg-tertiary) 50%, var(--bg-secondary) 100%);
-          border: 1px solid var(--border-subtle);
-          border-radius: var(--radius-lg);
-          padding: 32px;
-          display: flex;
-          flex-direction: column;
-          align-items: center;
-        }
-        .today-hero-skeleton {
-          display: flex;
-          flex-direction: column;
-          align-items: center;
-          padding: 32px;
-          background: var(--bg-secondary);
-          border: 1px solid var(--border-subtle);
-          border-radius: var(--radius-lg);
-        }
-        .today-hero-alert {
-          margin-top: 20px;
-          padding: 10px 16px;
-          border-radius: var(--radius-md);
-          font-size: 13px;
-          font-weight: 500;
-          animation: pulse 2s ease-in-out infinite;
-          background: rgba(248, 113, 113, 0.1);
-          border: 1px solid var(--error);
-          color: var(--error);
-        }
-        @keyframes pulse {
-          0%, 100% { opacity: 1; }
-          50% { opacity: 0.7; }
-        }
-
-        /* ── Ring Gauge ── */
-        .today-ring-wrap {
-          display: flex;
-          flex-direction: column;
-          align-items: center;
-          gap: 20px;
-        }
-        .today-ring-svg {
-          width: 180px;
-          height: 180px;
-          filter: drop-shadow(0 0 20px rgba(240, 160, 48, 0.15));
-        }
-        .today-ring-pct {
-          font-weight: 800;
-          font-size: 22px;
-          letter-spacing: -0.02em;
-        }
-        .today-ring-label {
-          font-size: 10px;
-          fill: var(--text-muted);
-          text-transform: uppercase;
-          letter-spacing: 0.1em;
-          font-weight: 600;
-        }
-        .today-ring-meta {
-          display: flex;
-          gap: 24px;
-          align-items: center;
-        }
-        .today-ring-stat {
-          display: flex;
-          flex-direction: column;
-          align-items: center;
-          gap: 4px;
-        }
-        .today-ring-stat-label {
-          font-size: 10px;
-          font-weight: 600;
-          text-transform: uppercase;
-          letter-spacing: 0.06em;
-          color: var(--text-muted);
-        }
-        .today-ring-stat-value {
-          font-size: 22px;
-          font-weight: 700;
-          letter-spacing: -0.02em;
-        }
-        .today-ring-divider {
-          width: 1px;
-          height: 32px;
-          background: var(--border);
-        }
-
-        /* ── Date & Updated ── */
-        .today-date {
-          font-size: 12px;
-          color: var(--text-muted);
-          margin-left: 12px;
-          font-weight: 400;
-        }
-        .today-utc-badge {
-          font-size: 11px;
-          font-weight: 600;
-          color: var(--accent);
-          background: rgba(240, 160, 48, 0.12);
-          padding: 2px 6px;
-          border-radius: 4px;
-          vertical-align: middle;
-          letter-spacing: 0.06em;
-          margin-left: 4px;
-        }
-        .today-utc-clock {
-          display: block;
-          font-size: 11px;
-          color: var(--text-muted);
-          margin-left: 12px;
-          margin-top: 2px;
-          font-variant-numeric: tabular-nums;
-        }
-        .today-updated {
-          font-size: 11px;
-          color: var(--text-muted);
-          padding: 4px 10px;
-          background: var(--bg-tertiary);
-          border-radius: var(--radius-sm);
-        }
-
-        /* ── No Budget ── */
-        .today-no-budget {
-          display: flex;
-          flex-direction: column;
-          align-items: center;
-          gap: 8px;
-          padding: 24px;
-        }
-        .today-no-budget-icon {
-          font-size: 40px;
-          opacity: 0.5;
-        }
-        .today-no-budget-title {
-          font-size: 16px;
-          font-weight: 600;
-        }
-        .today-no-budget-desc {
-          font-size: 13px;
-          color: var(--text-muted);
-          text-align: center;
-          max-width: 320px;
-        }
-
-        /* ── Stat Cards ── */
-        .today-stat-card {
-          background: linear-gradient(135deg, var(--bg-secondary) 0%, var(--bg-tertiary) 100%);
-          position: relative;
-          overflow: hidden;
-        }
-        .today-stat-card::before {
-          content: '';
-          position: absolute;
-          top: 0;
-          left: 0;
-          right: 0;
-          height: 2px;
-          background: linear-gradient(90deg, var(--accent), transparent);
-          opacity: 0;
-          transition: opacity 0.3s;
-        }
-        .today-stat-card:hover::before {
-          opacity: 1;
-        }
-
-        /* ── Model Token Breakdown ── */
-        .today-model-list {
-          padding: 8px 0;
-        }
-        .today-model-row {
-          padding: 12px 20px;
-          border-bottom: 1px solid var(--border-subtle);
-          transition: background 0.15s;
-        }
-        .today-model-row:last-child {
-          border-bottom: none;
-        }
-        .today-model-row:hover {
-          background: var(--bg-hover);
-        }
-        .today-model-info {
-          display: flex;
-          align-items: baseline;
-          gap: 8px;
-          margin-bottom: 6px;
-        }
-        .today-model-name {
-          font-size: 13px;
-          font-weight: 600;
-          color: var(--text-primary);
-        }
-        .today-model-provider {
-          font-size: 11px;
-          color: var(--text-muted);
-          padding: 1px 6px;
-          background: var(--bg-tertiary);
-          border-radius: 4px;
-        }
-        .today-model-metrics {
-          display: flex;
-          gap: 20px;
-          margin-bottom: 8px;
-        }
-        .today-model-metric {
-          display: flex;
-          flex-direction: column;
-          gap: 1px;
-        }
-        .today-model-metric-label {
-          font-size: 10px;
-          font-weight: 600;
-          text-transform: uppercase;
-          letter-spacing: 0.04em;
-          color: var(--text-muted);
-        }
-        .today-model-metric-value {
-          font-size: 13px;
-          font-weight: 600;
-          color: var(--text-primary);
-          font-variant-numeric: tabular-nums;
-        }
-
-        /* ── Token bar ── */
-        .today-token-bar {
-          display: flex;
-          height: 6px;
-          border-radius: 3px;
-          overflow: hidden;
-          background: var(--bg-tertiary);
-        }
-        .today-token-bar-in {
-          background: var(--info);
-          transition: width 0.6s ease;
-        }
-        .today-token-bar-out {
-          background: var(--accent);
-          transition: width 0.6s ease;
-        }
-
-        /* ── Legend ── */
-        .today-token-legend {
-          display: flex;
-          gap: 16px;
-          align-items: center;
-        }
-        .today-legend-item {
-          display: flex;
-          align-items: center;
-          gap: 6px;
-          font-size: 11px;
-          color: var(--text-muted);
-          font-weight: 500;
-        }
-        .today-legend-dot {
-          width: 8px;
-          height: 8px;
-          border-radius: 2px;
-        }
-        .today-legend-in {
-          background: var(--info);
-        }
-        .today-legend-out {
-          background: var(--accent);
-        }
-
-        /* ── Skeletons ── */
-        .skeleton {
-          background: linear-gradient(90deg, var(--bg-tertiary) 25%, var(--bg-elevated) 50%, var(--bg-tertiary) 75%);
-          background-size: 200% 100%;
-          animation: shimmer 1.5s infinite;
-          border-radius: var(--radius-md);
-        }
-        @keyframes shimmer {
-          0% { background-position: 200% 0; }
-          100% { background-position: -200% 0; }
-        }
-
-        /* ── Grant Cards ── */
-        .today-grant-list {
-          padding: 8px 0;
-        }
-        .today-grant-card {
-          padding: 16px 20px;
-          border-bottom: 1px solid var(--border-subtle);
-          transition: background 0.15s;
-        }
-        .today-grant-card:last-child {
-          border-bottom: none;
-        }
-        .today-grant-card:hover {
-          background: var(--bg-hover);
-        }
-        .today-grant-header {
-          display: flex;
-          justify-content: space-between;
-          align-items: baseline;
-          margin-bottom: 10px;
-        }
-        .today-grant-reason {
-          font-size: 13px;
-          font-weight: 600;
-          color: var(--text-primary);
-        }
-        .today-grant-expiry {
-          font-size: 11px;
-          color: var(--text-muted);
-          padding: 2px 8px;
-          background: var(--bg-tertiary);
-          border-radius: 4px;
-        }
-        .today-grant-bar-wrap {
-          height: 6px;
-          background: var(--bg-tertiary);
-          border-radius: 3px;
-          overflow: hidden;
-          margin-bottom: 10px;
-        }
-        .today-grant-bar-fill {
-          height: 100%;
-          border-radius: 3px;
-          transition: width 0.8s cubic-bezier(0.4, 0, 0.2, 1);
-        }
-        .today-grant-stats {
-          display: flex;
-          gap: 20px;
-        }
-        .today-grant-stat {
-          display: flex;
-          flex-direction: column;
-          gap: 1px;
-        }
-        .today-grant-stat-label {
-          font-size: 10px;
-          font-weight: 600;
-          text-transform: uppercase;
-          letter-spacing: 0.04em;
-          color: var(--text-muted);
-        }
-        .today-grant-stat-value {
-          font-size: 13px;
-          font-weight: 600;
-          color: var(--text-primary);
-          font-variant-numeric: tabular-nums;
-        }
-        .today-grant-by {
-          font-weight: 500;
-          color: var(--text-muted);
-        }
-        .today-grant-count {
-          font-size: 11px;
-          color: var(--text-muted);
-          padding: 2px 8px;
-          background: var(--bg-tertiary);
-          border-radius: 4px;
-        }
-      `}</style>
     </>
   );
 }
