+++ ui/src/components/today/utils.ts
@@ -1,0 +1,6 @@
+/** Format a number with k/M suffixes for large token counts. */
+export function fmtTokens(n: number): string {
+  if (n >= 1_000_000) return `${(n / 1_000_000).toFixed(1)}M`;
+  if (n >= 1_000) return `${(n / 1_000).toFixed(1)}k`;
+  return n.toLocaleString();
+}
